agents/agent-development-kit/features/usage/requirements.txt
google-adk>=1.0.0
python-dotenv
google-generativeai